fin con carga por url

+++ MVT/app_familia/models.py
@@ -6,6 +6,6 @@ class Familiar(models.Model):
     nombre = models.CharField(max_length=30)
     apellido = models.CharField(max_length=30)
     email = models.EmailField()
-    edad = models.BigIntegerField()
+    edad = models.IntegerField()
     anioNacimiento = models.DateField()
